% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-global.R
\docType{data}
\name{array}
\alias{array}
\title{Sample ExprsBinary Data}
\format{An object of class \code{ExprsBinary} of length 1.}
\usage{
data(array)
}
\description{
Sample ExprsBinary Data
}
\keyword{datasets}
