% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1-import.R
\name{arrayEset}
\alias{arrayEset}
\title{Import Data from eSet}
\usage{
arrayEset(eSet, colBy, include)
}
\arguments{
\item{eSet}{An \code{eSet} object.}

\item{colBy}{A numeric or character index. The column that contains group annotations.}

\item{include}{A list of character vectors. Specifies which annotations in \code{colBy}
to include into which groups. Each element of a list specifies a unique group while
each element of the character vector specifies an annotation to fit to that group. For
binary classification, the first list element defines the negative or control group.}
}
\description{
A convenience function that builds an \code{ExprsArray} object from an \code{eSet} object.
}
\details{
The package Biobase maintains a popular class object called \code{ExpressionSet} that
 often gets used to store expression data. This function converts this \code{eSet}
 object into an \code{ExprsArray} object. This function automatically removes any
 features with \code{NA} values.
}
\seealso{
\code{\link{ExprsArray-class}}, \code{\link{arrayRead}}, \code{\link{GSE2eSet}}
}

