% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-global.R
\docType{data}
\name{arrayMulti}
\alias{arrayMulti}
\title{Sample ExprsMulti Data}
\format{An object of class \code{ExprsMulti} of length 1.}
\usage{
data(arrayMulti)
}
\description{
Sample ExprsMulti Data
}
\keyword{datasets}
