% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1-import.R
\name{arrayRead}
\alias{arrayRead}
\title{Import Data from File}
\usage{
arrayRead(file, probes.begin, colID, colBy, include, ...)
}
\arguments{
\item{file}{A character string. Argument passed along to \code{read.delim}.}

\item{probes.begin}{A numeric scalar. The j-th column at which feature data begins.}

\item{colID}{A numeric or character index. The column by which to name subjects.}

\item{colBy}{A numeric or character index. The column that contains group annotations.}

\item{include}{A list of character vectors. Specifies which annotations in \code{colBy}
to include into which groups. Each element of a list specifies a unique group while
each element of the character vector specifies an annotation to fit to that group. For
binary classification, the first list element defines the negative or control group.}
}
\description{
A convenience function to build an \code{ExprsArray} object from a tab-delimited file.
}
\details{
\code{arrayRead} conveniently builds an \code{ExprsArray} object from a tab-delimited
 data file, passing along the \code{file} and \code{...} argument(s) to
 \code{\link{read.delim}}. This function expects the delimited data file to assume
 the following format: rows indicate subject entries while columns indicate measured
 variables. The first several columns should contain annotation information (e.g. age,
 sex, diagnosis). The remaining columns should contain the feature (e.g. expression)
 data. The argument \code{probes.begin} defines the j-th column at which feature data
 begins. By default, \code{arrayRead} forces \code{stringsAsFactors = FASE}. This
 function automatically removes probes with missing (i.e. \code{NA}) values.
}
\seealso{
\code{\link{ExprsArray-class}}, \code{\link{arrayEset}}, \code{\link{GSE2eSet}}
}

