% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2-build.R
\name{buildANN}
\alias{buildANN}
\title{Build Artificial Neural Network Model}
\usage{
buildANN(object, top = 0, ...)
}
\arguments{
\item{object}{An \code{ExprsArray} object. The training set.}

\item{top}{A numeric scalar or character vector. A numeric scalar indicates
the number of top features that should undergo feature selection. A character vector
indicates specifically which features by name should undergo feature selection.
Set \code{top = 0} to include all features. A numeric vector can also be used
to indicate specific features by location, similar to a character vector.}

\item{...}{Arguments passed to the detailed function.}
}
\value{
Returns an \code{ExprsModel} object.
}
\description{
\code{buildANN} builds a model using the \code{nnet} function
 from the \code{nnet} package.
}
