% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.4-plMonteCarlo.R
\name{calcMonteCarlo}
\alias{calcMonteCarlo}
\title{Calculate \code{plMonteCarlo} Performance}
\usage{
calcMonteCarlo(pl, colBy = "valid.acc")
}
\arguments{
\item{pl}{Specifies the \code{ExprsPipeline} object returned by \code{plMonteCarlo}.}

\item{colBy}{A character vector or string. Specifies column(s) to use when
summarizing model performance. Listing multiple columns will calculate
performance as a product of those listed performances.}
}
\value{
A numeric scalar. The cross-validation accuracy.
}
\description{
\code{calcMonteCarlo} calculates a single performance measure for the
 results of a \code{plMonteCarlo} function call.
}
\details{
For each dataset split (i.e., bootstrap), \code{calcMonteCarlo} averages
 the validation set performance for the "best" model (where "best" is
 defined as the model with the maximum "internal" cross-validation
 accuracy, \code{max($train.plCV)}). The validation set performance
 ultimately averaged depends on the supplied \code{colBy} argument.
}
