% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.4-plNested.R
\name{check.ctrlGS}
\alias{check.ctrlGS}
\title{Check \code{ctrlGS} Arguments}
\usage{
check.ctrlGS(args)
}
\arguments{
\item{args}{A list of arguments to check.}
}
\description{
This function ensures that a list of arguments for \code{ctrlGS} meets the
 criteria for the \code{\link{plNested}} function.
}
\details{
This function mandates a non-NULL \code{fold} argument and a \code{TRUE}
 \code{aucSkip} argument.
}

