% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-global.R
\name{classCheck}
\alias{classCheck}
\title{Class Check}
\usage{
classCheck(x, what, msg)
}
\arguments{
\item{x}{An object.}

\item{what}{A character vector. The classes any of which \code{x} should have.}

\item{msg}{A string. An error message if \code{x} is not \code{what}.}
}
\description{
Checks whether an object belongs to a specified class.
 For back-end use only.
}
