% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-conjoin.R
\docType{methods}
\name{conjoin}
\alias{conjoin}
\alias{conjoin,ExprsArray-method}
\alias{conjoin,ExprsModel-method}
\alias{conjoin,ExprsPipeline-method}
\alias{conjoin,ExprsEnsemble-method}
\title{Combine \code{exprso} Objects}
\usage{
conjoin(object, ...)

\S4method{conjoin}{ExprsArray}(object, ...)

\S4method{conjoin}{ExprsModel}(object, ...)

\S4method{conjoin}{ExprsPipeline}(object, ...)

\S4method{conjoin}{ExprsEnsemble}(object, ...)
}
\arguments{
\item{object}{Any \code{exprso} object.}

\item{...}{More objects of the same class.}
}
\value{
See Details.
}
\description{
\code{conjoin} combines two or more \code{exprso} objects based on their class.
}
\details{
When joining \code{ExprsArray} objects, this function returns one
 \code{ExprsArray} object as output. This only works on \code{ExprsArray} objects
 that have not undergone feature selection. Any missing annotations in \code{@annot}
 will get replaced with \code{NA} values.

When joining \code{ExprsModel} or \code{ExprsEnsemble} objects,
 this function returns an ensemble.

When joining \code{ExprsPipeline} objects, this function returns one
 \code{ExprsPipeline} object as output. To track which \code{ExprsPipeline}
 objects contributed to the resultant object, the source gets flagged
 with a \code{boot} column. If a pipeline already has a \code{boot} column,
 the original boot tracker will receive an offset (and the old \code{boot}
 column will get renamed to \code{unboot}). This system ensures that all
 models deriving from the same training set will get handled as a
 "pseudo-bootstrap" by downstream \code{\link{pipe}} functions.
}
\section{Methods (by class)}{
\itemize{
\item \code{ExprsArray}: Method to join \code{ExprsArray} objects.

\item \code{ExprsModel}: Method to join \code{ExprsModel} objects.

\item \code{ExprsPipeline}: Method to join \code{ExprsPipeline} objects.

\item \code{ExprsEnsemble}: Method to join \code{ExprsEnsemble} objects.
}}

