% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{ctrlFeatureSelect}
\alias{ctrlFeatureSelect}
\title{Manage \code{fs} Arguments}
\usage{
ctrlFeatureSelect(func, top = 0, ...)
}
\arguments{
\item{func}{A character string. The \code{fs} function to call.}

\item{top}{Argument passed to the \code{fs} function.}

\item{...}{Additional arguments passed to the \code{fs} function.}
}
\value{
A list of arguments.
}
\description{
This function organizes \code{fs} arguments passed to \code{pl} functions.
}
