% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.3-plMonteCarlo.R
\name{ctrlGridSearch}
\alias{ctrlGridSearch}
\title{Manage \code{plGrid} Arguments}
\usage{
ctrlGridSearch(func, top, ...)
}
\arguments{
\item{func}{A character string. The \code{pl} function to call.}

\item{top}{Argument passed to the \code{pl} function.}

\item{...}{Additional arguments passed to the \code{pl} function.}
}
\value{
A list of arguments.
}
\description{
This function organizes \code{plGrid} arguments passed to \code{pl} functions.
}
\details{
\code{\link{plMonteCarlo}} and \code{\link{plNested}} functions currently
 only accept \code{\link{plGrid}} as the \code{ctrlGridSearch} \code{func}.
}

