% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{ctrlModSet}
\alias{ctrlModSet}
\title{Manage \code{mod} Arguments}
\usage{
ctrlModSet(func, ...)
}
\arguments{
\item{func}{A character string. The \code{mod} function to call.}

\item{...}{Additional arguments passed to the \code{mod} function.}
}
\value{
A list of arguments.
}
\description{
This function organizes \code{mod} arguments passed to \code{pl} functions.
}
