% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.4-plMonteCarlo.R
\name{ctrlSplitSet}
\alias{ctrlSplitSet}
\title{Manage \code{split} Arguments}
\usage{
ctrlSplitSet(func, percent.include, ...)
}
\arguments{
\item{func}{A character string. The \code{split} function to call.}

\item{percent.include}{Argument passed to the \code{split} function.}

\item{...}{Additional arguments passed to the \code{split} function.}
}
\value{
A list of arguments.
}
\description{
This function organizes \code{split} arguments passed to \code{pl} functions.
}
