% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-global.R
\name{defaultArg}
\alias{defaultArg}
\title{Set an args List Element to Default Value}
\usage{
defaultArg(what, as, args)
}
\arguments{
\item{what}{The name of the argument.}

\item{as}{The value to set it as.}

\item{args}{An args list. The result of \code{\link{getArgs}}.}
}
\description{
Set an args List Element to Default Value
}
