% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1-classes.R, R/1.2-methods.R
\docType{class}
\name{ExprsArray-class}
\alias{$,ExprsArray-method}
\alias{ExprsArray-class}
\alias{[,ExprsArray,ANY,ANY,ANY-method}
\alias{getProbeSet,ExprsArray-method}
\alias{plot,ExprsArray,missing-method}
\alias{show,ExprsArray-method}
\alias{subset,ExprsArray-method}
\alias{summary,ExprsArray-method}
\title{An S4 class to store feature and annotation data}
\usage{
\S4method{show}{ExprsArray}(object)

\S4method{[}{ExprsArray,ANY,ANY,ANY}(x, i, j, drop)

\S4method{$}{ExprsArray}(x, name)

\S4method{subset}{ExprsArray}(x, subset, select)

\S4method{plot}{ExprsArray,missing}(x, i = 1, j = 2, k = 3, colors,
  shapes)

\S4method{summary}{ExprsArray}(object)

\S4method{getProbeSet}{ExprsArray}(object)
}
\arguments{
\item{object, x}{An object of class \code{ExprsArray}.}

\item{i, j, drop}{Subsets via \code{object@annot[i, j, drop]}.}

\item{name}{Subsets via \code{object@annot[, name]}.}

\item{subset}{Subsets via \code{object@annot[subset, ]}.
Use this argument to rearrange feature order.}

\item{select}{Subsets via \code{object@annot[, select]}.
Use this argument to rearrange subject order.}

\item{colors}{A character vector. Optional. Manually assign a color to each subject point.}

\item{shapes}{A numeric vector. Optional. Manually assign a shape to each subject point.}

\item{i, j, k}{A numeric scalar. Indexes the first, second, and third dimensions to plot.}
}
\description{
An S4 class to store feature and annotation data
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{ExprsArray} object.

\item \code{[}: Method to subset \code{ExprsArray} object.

\item \code{$}: Method to subset \code{ExprsArray} object.

\item \code{subset}: Method to subset \code{ExprsArray} object.

\item \code{plot}: Method to plot three dimensions of the expression data.

\item \code{summary}: Method to plot summary graphs for a sub-sample of expression data.

\item \code{getProbeSet}: Method to return features within an \code{ExprsArray} object.
}}
\section{Slots}{

\describe{
\item{\code{exprs}}{A matrix. Stores the expression data.}

\item{\code{annot}}{A data.frame. Stores the annotation data.}

\item{\code{preFilter}}{Typically a list. Stores feature selection history.}

\item{\code{reductionModel}}{Typically a list. Stores dimension reduction history.}
}}
\seealso{
\code{\link{ExprsArray-class}}\cr
\code{\link{ExprsModel-class}}\cr
\code{\link{ExprsPipeline-class}}\cr
\code{\link{ExprsEnsemble-class}}\cr
\code{\link{ExprsPredict-class}}
}

