% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1-classes.R, R/1.2-methods.R
\docType{class}
\name{ExprsEnsemble-class}
\alias{ExprsEnsemble-class}
\alias{show,ExprsEnsemble-method}
\alias{getFeatures,ExprsEnsemble-method}
\title{An S4 class to store multiple classification models}
\usage{
\S4method{show}{ExprsEnsemble}(object)

\S4method{getFeatures}{ExprsEnsemble}(object, index)
}
\arguments{
\item{object}{An \code{ExprsArray}, \code{ExprsModel}, \code{ExprsPipeline},
or \code{ExprsEnsemble} object.}

\item{index}{A numeric scalar. The i-th model from which to retrieve features.
If missing, \code{getFeatures} will tabulate features across all models.}
}
\description{
An S4 class to store multiple classification models
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{ExprsEnsemble} object.

\item \code{getFeatures}: Method to return features within an \code{ExprsEnsemble} model.
}}

\section{Slots}{

\describe{
\item{\code{machs}}{Typically a list. Stores the classification models.}
}}

\seealso{
\code{\link{ExprsArray-class}}\cr
\code{\link{ExprsModel-class}}\cr
\code{\link{ExprsPipeline-class}}\cr
\code{\link{ExprsEnsemble-class}}\cr
\code{\link{ExprsPredict-class}}
}
