% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1-classes.R
\docType{class}
\name{ExprsMulti-class}
\alias{ExprsMulti-class}
\title{An S4 class to store feature and annotation data}
\description{
An \code{ExprsArray} sub-class for data with multiple class outcomes.
}
