% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-predict.R
\docType{methods}
\name{exprso-predict}
\alias{exprso-predict}
\alias{predict,ExprsMachine-method}
\alias{predict,ExprsModule-method}
\title{Predict Class Labels}
\usage{
\S4method{predict}{ExprsMachine}(object, array, verbose = TRUE)

\S4method{predict}{ExprsModule}(object, array, verbose = TRUE)
}
\arguments{
\item{object}{An \code{ExprsModel} object. The classifier to deploy.}

\item{array}{An \code{ExprsArray} object. The validation set.}

\item{verbose}{A logical scalar. Toggles whether to print \code{calcStats}.}
}
\value{
Returns an \code{ExprsPredict} object.
}
\description{
Predict class labels of a validation set.
}
\details{
An \code{ExprsMachine} object can only predict against an \code{ExprsBinary}
 object. An \code{ExprsModule} object can only predict against an
 \code{ExprsMulti} object. The validation set should never get modified
 once separated from the training set. If the training set used to build
 an \code{ExprsModule} had a class missing (i.e. a NULL placeholder),
 the \code{ExprsModule} cannot predict the missing class. As with
 all functions included in this package, ties get resolved using
 probability weights based on the relative frequency of the tied
 classes.

\code{ExprsPredict} objects store predictions in three slots: \code{@pred},
 \code{@decision.values}, and \code{@probability}. The first slot
 stores a "final decision" based on the class label with the maximum
 predicted probability. The second slot stores a transformation
 of the predicted probability for each class calculated by the inverse
 of Platt scaling. The predicted probability gets returned by the
 \code{predict} method called using the stored \code{@mach} object.
 To learn how these slots get used to calculate classifier performance,
 read more at \code{\link{calcStats}}.
}
\seealso{
\code{\link{modHistory}}, \code{\link{calcStats}}
}

