% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.1-import.R
\name{exprso}
\alias{exprso}
\title{Import Data as ExprsArray}
\usage{
exprso(x, y)
}
\arguments{
\item{x}{A matrix of feature data for all samples. Rows should
contain samples and columns should contain features.}

\item{y}{A vector of outcomes for all samples. If
\code{class(y) == "character"}, \code{exprso} prepares data
for binary or multi-class classification. Else, \code{exprso}
prepares data for regression. If \code{y} is a matrix, the
program assumes the first column is the outcome.}
}
\value{
An \code{ExprsArray} object.
}
\description{
A convenience function that builds an \code{ExprsArray} object.
}
