% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1-classes.R, R/1.2-methods.R
\docType{class}
\name{ExprsPipeline-class}
\alias{$,ExprsPipeline-method}
\alias{ExprsPipeline-class}
\alias{[,ExprsPipeline,ANY,ANY,ANY-method}
\alias{getProbeSet,ExprsPipeline-method}
\alias{show,ExprsPipeline-method}
\alias{subset,ExprsPipeline-method}
\alias{summary,ExprsPipeline-method}
\title{An S4 class to store models built during high-throughput learning}
\usage{
\S4method{show}{ExprsPipeline}(object)

\S4method{[}{ExprsPipeline,ANY,ANY,ANY}(x, i, j, drop)

\S4method{$}{ExprsPipeline}(x, name)

\S4method{subset}{ExprsPipeline}(x, subset, select)

\S4method{summary}{ExprsPipeline}(object)

\S4method{getProbeSet}{ExprsPipeline}(object, index)
}
\arguments{
\item{object}{An object of class \code{ExprsPipeline}.}

\item{i, j, drop}{Subsets via \code{object@summary[i, j, drop]}.}

\item{name}{Subsets via \code{object@summary[, name]}.}

\item{subset}{Subsets via \code{object@summary[subset, ]}.
Use this argument to rearrange feature order.}

\item{select}{Subsets via \code{object@summary[, select]}.
Use this argument to rearrange subject order.}

\item{index}{A numeric scalar. Indicates the i-th model from which to retrieve the features. If missing,
\code{getProbeSet} will tabulate features across all models.}
}
\description{
An S4 class to store models built during high-throughput learning
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{ExprsPipeline} object.

\item \code{[}: Method to subset \code{ExprsPipeline} object.

\item \code{$}: Method to subset \code{ExprsPipeline} object.

\item \code{subset}: Method to subset \code{ExprsPipeline} object.

\item \code{summary}: Method to summarize \code{ExprsPipeline} parameters and performances.

\item \code{getProbeSet}: Method to return features within an \code{ExprsPredict} model.
}}
\section{Slots}{

\describe{
\item{\code{summary}}{Typically a data.frame. Stores the parameters and performances for classification models.}

\item{\code{machs}}{Typically a list. Stores the classification models referenced in \code{summary} slot.}
}}
\seealso{
\code{\link{ExprsArray-class}}\cr
\code{\link{ExprsModel-class}}\cr
\code{\link{ExprsPipeline-class}}\cr
\code{\link{ExprsEnsemble-class}}\cr
\code{\link{ExprsPredict-class}}
}

