% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1-classes.R, R/1.2-methods.R
\docType{class}
\name{ExprsPredict-class}
\alias{ExprsPredict-class}
\alias{show,ExprsPredict-method}
\title{An S4 class to store class predictions}
\usage{
\S4method{show}{ExprsPredict}(object)
}
\arguments{
\item{object}{An object of class \code{ExprsPredict}.}
}
\description{
An S4 class to store class predictions
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{ExprsPredict} object.
}}
\section{Slots}{

\describe{
\item{\code{pred}}{A factor. Stores class predictions as an unambiguous
class assignment.}

\item{\code{decision.values}}{Typically a matrix. Stores class predictions
as a decision value.}

\item{\code{probability}}{Typically a matrix. Stores class predictions
as a probability.}

\item{\code{actual}}{Typically a factor. Stores known class labels.
Used by \code{\link{calcStats}}.}
}}
\seealso{
\code{\link{ExprsArray-class}}\cr
\code{\link{ExprsModel-class}}\cr
\code{\link{ExprsPipeline-class}}\cr
\code{\link{ExprsEnsemble-class}}\cr
\code{\link{ExprsPredict-class}}
}

