% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1-fs.R
\name{fsAnnot}
\alias{fsAnnot}
\title{Use Annotations as Features}
\usage{
fsAnnot(object, colBy = NULL)
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo feature selection.}

\item{colBy}{A character vector. The names of annotations to rank above all others.}
}
\value{
Returns an \code{ExprsArray} object.
}
\description{
\code{fsAnnot} moves annotations named by the \code{colBy} argument
 to the top of the ranked features list. Otherwise, the relative order
 of the features does not change.
}
