% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1-fs.R
\name{fsPRA}
\alias{fsPRA}
\title{Reduce Dimensions by Log-Ratio Selection}
\usage{
fsPRA(object, top = 0, ...)
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo feature selection.}

\item{top}{A numeric scalar or character vector. A numeric scalar indicates
the number of top features that should undergo feature selection. A character vector
indicates specifically which features by name should undergo feature selection.
Set \code{top = 0} to include all features. A numeric vector can also be used
to indicate specific features by location, similar to a character vector.}

\item{...}{Arguments passed to the detailed function.}
}
\value{
Returns an \code{ExprsArray} object.
}
\description{
\code{fsPRA} finds the most explanatory pairwise log-ratios
 using the variable selection method proposed by Michael Greenacre
 in "Variable Selection in Compositional Data Analysis Using
 Pairwise Logratios", modified to run faster.
}
