% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1-fs.R
\name{fsPrcomp}
\alias{fsPrcomp}
\title{Reduce Dimensions by PCA}
\usage{
fsPrcomp(object, top = 0, ...)
}
\arguments{
\item{object}{Specifies the \code{ExprsArray} object to undergo feature selection.}

\item{top}{A numeric scalar or character vector. A numeric scalar indicates
the number of top features that should undergo feature selection. A character vector
indicates specifically which features by name should undergo feature selection.
Set \code{top = 0} to include all features. A numeric vector can also be used
to indicate specific features by location, similar to a character vector.}

\item{...}{Arguments passed to the detailed function.}
}
\value{
Returns an \code{ExprsArray} object.
}
\description{
\code{fsPrcomp} reduces dimensions using the \code{prcomp} function.
 The reduction model is saved and deployed automatically on any new
 data during model validation.
}
