% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1-fs.R
\name{fsPropd}
\alias{fsPropd}
\title{Select Features by Differential Proportionality Analysis}
\usage{
fsPropd(object, top = 0, keep = 0, modRatios = FALSE, ...)
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo feature selection.}

\item{top}{A numeric scalar or character vector. A numeric scalar indicates
the number of top features that should undergo feature selection. A character vector
indicates specifically which features by name should undergo feature selection.
Set \code{top = 0} to include all features. A numeric vector can also be used
to indicate specific features by location, similar to a character vector.}

\item{keep}{A numeric scalar. Specifies the number of top features that should get
returned by the feature selection method. Use of \code{keep} is generally not
recommended, but can speed up analyses of large data.}

\item{modRatios}{A logical scalar. Toggles whether to compute theta from
the feature ratios as provided. Set \code{modRatios = TRUE} if data were
recasted by a prior \code{modRatios} call. If \code{TRUE}, the \code{alpha}
and \code{weighted} arguments will not work.}

\item{...}{Arguments passed to the detailed function.}
}
\value{
Returns an \code{ExprsArray} object.
}
\description{
\code{fsPropd} selects features using the \code{propd} function
 from the \code{propr} package.
}
