% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2-methods.R
\name{getProbeSet}
\alias{getProbeSet}
\title{Retrieve \code{exprso} Feature Set}
\usage{
getProbeSet(object, ...)
}
\arguments{
\item{object}{An \code{ExprsArray}, \code{ExprsModel}, \code{ExprsPipeline},
\code{ExprsEnsemble}, or \code{ExprsPredict} object.}

\item{...}{See \code{\link{ExprsPipeline-class}} or
\code{\link{ExprsEnsemble-class}}.}
}
\description{
See the respective S4 class for method details.
}
\seealso{
\code{\link{ExprsArray-class}}\cr
\code{\link{ExprsModel-class}}\cr
\code{\link{ExprsPipeline-class}}\cr
\code{\link{ExprsEnsemble-class}}\cr
\code{\link{ExprsPredict-class}}
}

