% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2-methods.R
\name{getWeights}
\alias{getWeights}
\title{Retrieve LASSO Weights}
\usage{
getWeights(object, ...)
}
\arguments{
\item{object}{An \code{ExprsModel}, \code{ExprsPipeline},
or \code{ExprsEnsemble} object.}

\item{...}{Arguments passed to \code{glmnet::coef.cv.glmnet}.}
}
\description{
See the respective S4 class for method details.
}
