% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\docType{data}
\name{Golub_Merge}
\alias{Golub_Merge}
\title{Golub Data Copy}
\format{An object of class \code{ExpressionSet} with 7129 rows and 72 columns.}
\usage{
data(Golub_Merge)
}
\description{
Golub Data Copy
}
\keyword{datasets}
