% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{lequal}
\alias{lequal}
\title{Test All Equal Within List}
\usage{
lequal(list)
}
\arguments{
\item{list}{A list.}
}
\description{
This function tests whether all elements in a list are identical.
 Works like an iterative \code{all.equal}.
}
