% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.2-plGrid.R
\name{makeGridFromArgs}
\alias{makeGridFromArgs}
\title{Build Argument Grid}
\usage{
makeGridFromArgs(array.train, top, how, ...)
}
\arguments{
\item{array.train}{The \code{array.train} argument as fed to \code{plGrid}.}

\item{top}{The \code{top} argument as fed to \code{plGrid}.}

\item{how}{The \code{how} argument as fed to \code{plGrid}.}

\item{...}{Additional arguments as fed to \code{plGrid}.}
}
\description{
This function builds an argument grid from any number of arguments.
 Used to prepare a grid-search for the \code{plGrid} and
 \code{plGridMulti} functions.
}
