% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-mod.R
\name{modAcomp}
\alias{modAcomp}
\title{Compositionally Constrain Data}
\usage{
modAcomp(object)
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo pre-processing.}
}
\value{
A pre-processed \code{ExprsArray} object.
}
\description{
\code{modAcomp} makes it so that all sample vectors have the same total sum.
}
