% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-mod.R
\name{modCLR}
\alias{modCLR}
\title{Log-ratio Transform Data}
\usage{
modCLR(object)
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo pre-processing.}
}
\value{
An \code{ExprsArray} object.
}
\description{
\code{modCLR} applies a centered log-ratio transformation to the data.
}
