% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-mod.R
\name{modHistory}
\alias{modHistory}
\title{Replicate Data Process History}
\usage{
modHistory(object, reference)
}
\arguments{
\item{object}{An \code{ExprsArray} object. The object that should undergo a
replication of the feature selection history.}

\item{reference}{An \code{ExprsArray} or \code{ExprsModel} object. The object
containing the reference history.}
}
\value{
An \code{ExprsArray} object.
}
\description{
\code{modHistory} replicates the feature selection history of a reference.
 Used by \code{predict} to prepare test set for model deployment.
}
