% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-mod.R
\name{modInclude}
\alias{modInclude}
\title{Select Features from Data}
\usage{
modInclude(object, include = rownames(object@exprs))
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo pre-processing.}

\item{include}{A character vector. The names of features to include.}
}
\value{
A pre-processed \code{ExprsArray} object.
}
\description{
\code{modSelect} selects specific features from a data set. Unlike
 \code{fsInclude}, this function does not update \code{@preFilter}
 and returns only those features stated by \code{include}.
}
