% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-mod.R
\name{modNormalize}
\alias{modNormalize}
\title{Normalize Data}
\usage{
modNormalize(object, MARGIN = c(1, 2))
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo pre-processing.}

\item{MARGIN}{A numeric vector. The margin by which to normalize.
Provide \code{MARGIN = 1} to normalize the feature vector.
Provide \code{MARGIN = 2} to normalize the subject vector.
Provide \code{MARGIN = c(1, 2)} to normalize by the subject vector
and then by the feature vector.}
}
\value{
A pre-processed \code{ExprsArray} object.
}
\description{
\code{modNormalize} normalizes feature data.
}
\details{
This method normalizes subject and/or feature vectors according to the
 formula \code{y = (x - mean(x)) / sd(x)}.
}
