% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-mod.R
\name{modPermute}
\alias{modPermute}
\title{Permute Features in Data}
\usage{
modPermute(object)
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo pre-processing.}
}
\value{
An \code{ExprsArray} object.
}
\description{
\code{modPermute} randomly shuffles each feature across
 all samples in the data. Using permuted data can establish
 a null model for testing the significance of prediction
 error estimates. This approach preserves the univariate
 distributions, but will change the multivariate
 distribution.
}
