% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-mod.R
\name{modSample}
\alias{modSample}
\title{Sample Features from Data}
\usage{
modSample(object, size = 0)
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo pre-processing.}

\item{size}{A numeric scalar. The number of randomly sampled features
to include in the pre-processed \code{ExprsArray} object.}
}
\value{
A pre-processed \code{ExprsArray} object.
}
\description{
\code{modSample} samples features from a data set randomly without
 replacement. When \code{size = 0}, this is equivalent to
 \code{fsSample, top = 0}, but much quicker.
}
