% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-mod.R
\name{modScale}
\alias{modScale}
\title{Rescale Data by Factor Range}
\usage{
modScale(object, to = 2)
}
\arguments{
\item{object}{An \code{ExprsArray} object. The object that should undergo a
replication of some feature selection and dimension reduction history.}
}
\value{
A pre-processed \code{ExprsArray} object.
}
\description{
\code{modScale} rescales a data set by making all sample vectors
 have the same total sum, then multiplying each sample vector by
 a scaling factor. The scaling factor is randomly sampled from
 a uniform sequence ranging from 1 to the \code{to} argument.
}
