% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-mod.R
\name{modShuffle}
\alias{modShuffle}
\title{Shuffle Feature Order}
\usage{
modShuffle(object)
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo pre-processing.}
}
\value{
A pre-processed \code{ExprsArray} object.
}
\description{
\code{modShuffle} shuffles feature order by randomly sampling
 features without replacement. Formally, this is equivalent
 to \code{fsSample, top = 0}, but much quicker.
}
