% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2-modSwap.R
\docType{methods}
\name{modSwap}
\alias{modSwap}
\alias{modSwap,ExprsBinary-method}
\title{Swap Case Subjects}
\usage{
modSwap(object, ...)

\S4method{modSwap}{ExprsBinary}(object, from, percent = 10)
}
\arguments{
\item{object}{An \code{ExprsBinary} object containing cases to swap with controls.}

\item{from}{An \code{ExprsBinary} object from which to draw the control subjects.}

\item{percent}{A numeric scalar. The percentage of subjects to mutate.}
}
\value{
An \code{ExprsBinary} object containing swapped subjects with an index
 appended to the \code{$mutated} column of the \code{@annot} slot.
}
\description{
This experimental function swaps a percentage of case subjects from one dataset with
 control subjects from another dataset.
}
\section{Methods (by class)}{
\itemize{
\item \code{ExprsBinary}: A method to swap \code{ExprsBinary} objects.
}}

