% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.2-misc.R
\docType{methods}
\name{modTransform}
\alias{modTransform}
\alias{modTransform,ExprsArray-method}
\title{Log Transform Data}
\usage{
modTransform(object, ...)

\S4method{modTransform}{ExprsArray}(object, plotSkip = TRUE)
}
\arguments{
\item{object}{Specifies the \code{ExprsArray} object to undergo pre-processing.}

\item{plotSkip}{A logical scalar. Toggles whether to produce side-by-side plots
of the before and after \code{ExprsArray} summary.}
}
\description{
\code{modTransform} log (base 2) transforms gene expression feature data.
}
\section{Methods (by class)}{
\itemize{
\item \code{ExprsArray}: Method to transform an \code{ExprsArray} object.
}}
\seealso{
\code{\link{modFilter}}, \code{\link{modTransform}}, \code{\link{modNormalize}}
}

