% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1-classes.R, R/1.2-methods.R
\docType{class}
\name{MultiPredict-class}
\alias{MultiPredict-class}
\alias{show,MultiPredict-method}
\title{An S4 class to store model predictions}
\usage{
\S4method{show}{MultiPredict}(object)
}
\arguments{
\item{object}{An object of class \code{MultiPredict}.}
}
\description{
An S4 class to store model predictions
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{MultiPredict} object.
}}

\section{Slots}{

\describe{
\item{\code{pred}}{Any. Stores predicted outcome.}

\item{\code{actual}}{Any. Stores actual outcome.
Used by \code{\link{calcStats}}.}
}}

\seealso{
\code{\link{ExprsArray-class}}\cr
\code{\link{ExprsModel-class}}\cr
\code{\link{ExprsPipeline-class}}\cr
\code{\link{ExprsEnsemble-class}}\cr
\code{\link{ExprsPredict-class}}\cr
\code{\link{MultiPredict-class}}\cr
\code{\link{RegrsPredict-class}}
}
