% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2-methods.R
\name{nfeats}
\alias{nfeats}
\title{Get Number of Features}
\usage{
nfeats(object)
}
\arguments{
\item{object}{An \code{ExprsArray} object.}
}
\description{
This function returns the number of features in an \code{ExprsArray} object.
}
