% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2-methods.R
\name{nsamps}
\alias{nsamps}
\title{Get Number of Samples}
\usage{
nsamps(object)
}
\arguments{
\item{object}{An \code{ExprsArray} object.}
}
\description{
This function returns the number of samples in an \code{ExprsArray} object.
}
