% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3-exprso.R
\name{pipe}
\alias{pipe}
\title{Process Pipelines}
\description{
The \code{exprso} package includes these pipeline process modules:

- \code{\link{pipeSubset}}

- \code{\link{pipeFilter}}

- \code{\link{pipeUnboot}}
}
