% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8-pipe.R
\name{pipeUnboot}
\alias{pipeUnboot}
\title{Rename "boot" Column}
\usage{
pipeUnboot(object)
}
\arguments{
\item{object}{An \code{\link{ExprsPipeline-class}} object.}
}
\value{
An \code{\link{ExprsPipeline-class}} object.
}
\description{
\code{pipeUnboot} renames the "boot" column summary to "unboot".
}
\details{
This method provides a convenient adjunct to \code{\link{pipeFilter}} owing to
 how \code{exprso} handles \code{ExprsPipeline} objects with a "boot" column.
}
