% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3-exprso.R
\name{pl}
\alias{pl}
\title{Deploy Pipeline}
\description{
The \code{exprso} package includes these automated pipeline modules:

- \code{\link{plCV}}

- \code{\link{plGrid}}

- \code{\link{plGridMulti}}

- \code{\link{plMonteCarlo}}

- \code{\link{plNested}}
}
