% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.4-plNested.R
\name{plNested}
\alias{plNested}
\title{Nested Cross-Validation}
\usage{
plNested(array, fold = 10, ctrlFS, ctrlGS, save = FALSE)
}
\arguments{
\item{array}{Specifies the \code{ExprsArray} object to undergo cross-validation.}

\item{fold}{A numeric scalar. Specifies the number of folds for cross-validation.
Set \code{fold = 0} to perform leave-one-out cross-validation.}

\item{ctrlFS}{A list of arguments handled by \code{\link{ctrlFeatureSelect}}.}

\item{ctrlGS}{Arguments handled by \code{\link{ctrlGridSearch}}.}

\item{save}{A logical scalar. Toggles whether to save each fold.}
}
\value{
An \code{\link{ExprsPredict-class}} object.
}
\description{
Perform nested cross-validation.
}
\details{
Analogous to how \code{\link{plGrid}} manages multiple \code{build} and
 \code{predict} tasks, \code{plNested} effectively manages multiple
 \code{plGrid} tasks.

Specifically, \code{plNested} segregates the data into v-folds,
 treating each fold as a validation set to a training set built from
 the remaining subjects. Then, some \code{fold} times, it performs all
 feature selection tasks (listed via \code{ctrlFS}) on each split
 training set, and executing \code{plGrid} (via \code{ctrlGS}) on the
 training set. The resultant \code{ExprsPipeline} object contains the
 "outer" fold performance metrics in the columns whose names begin with
 "valid" (as carried over by \code{plGrid}).

To perform multiple feature selection tasks, supply a list of multiple
 \code{\link{ctrlFeatureSelect}} argument wrappers to \code{ctrlFS}.
 To reduce the results of \code{plNested} to a single performance metric,
 feed the returned \code{ExprsPipeline} object through
 \code{\link{calcNested}}.

Note that \code{plGrid} uses \code{\link{plCV}} to calculate the "inner"
 cross-validation accuracy. Depending on the use case, may not represent
 the most appropriate choice. We hope future implementations will accomodate
 other \code{ctrlGS} functions beyond \code{plGrid}.

When calculating classifier performance, this function forces
 \code{aucSkip = TRUE} and \code{plotSkip = TRUE}, not unlike
 \code{\link{plCV}}.
}
\examples{
\dontrun{
require(golubEsets)
data(Golub_Merge)
array <- arrayEset(Golub_Merge, colBy = "ALL.AML", include = list("ALL", "AML"))
array <- modFilter(array, 20, 16000, 500, 5) # pre-filter Golub ala Deb 2003
array <- modTransform(array) # lg transform
array <- modNormalize(array, c(1, 2)) # normalize gene and subject vectors
fs <- ctrlFeatureSelect(func = "fsEbayes", probes = 0)
gs <- ctrlGridSearch(func = "plGrid", how = "buildANN", probes = c(10, 20, 30),
                     size = 1:3, decay = c(0, .5, 1), fold = 0)
nest <- plNested(arrays[[1]], fold = 10, ctrlFS = fs, ctrlGS = gs, save = FALSE)
}
}
\seealso{
\code{\link{plCV}}, \code{\link{plGrid}}, \code{\link{plMonteCarlo}}, \code{\link{plNested}}
}

