% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1-classes.R
\docType{class}
\name{RegrsPredict-class}
\alias{RegrsPredict-class}
\title{An S4 class to store model predictions}
\description{
An S4 class to store model predictions
}
\section{Slots}{

\describe{
\item{\code{pred}}{Any. Stores predicted outcome.}

\item{\code{actual}}{Any. Stores actual outcome.
Used by \code{\link{calcStats}}.}
}}

\seealso{
\code{\link{ExprsArray-class}}\cr
\code{\link{ExprsModel-class}}\cr
\code{\link{ExprsPipeline-class}}\cr
\code{\link{ExprsEnsemble-class}}\cr
\code{\link{ExprsPredict-class}}
}
