% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-split.R
\name{splitBalanced}
\alias{splitBalanced}
\title{Split by Balanced Sampling}
\usage{
splitBalanced(object, percent.include = 67, ...)
}
\arguments{
\item{object}{An \code{ExprsArray} object to split.}

\item{percent.include}{Specifies the percent of the total number
of subjects to include in the training set.}

\item{...}{Arguments passed to both \code{splitStratify} calls.}
}
\value{
Returns a list of two \code{ExprsArray} objects.
}
\description{
\code{splitBalance} is a wrapper that calls \code{splitStratify}
 twice. In the first call, \code{splitStratify} is used to create a
 balanced training set from the total data. In the second call,
 \code{splitStratify} is used to create a balanced validation set
 from the leftover data. This function ensures that there are always
 an equal number of samples from each class in the split.
}
