% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-split.R
\name{splitBoost}
\alias{splitBoost}
\title{Sample by Boosting}
\usage{
splitBoost(object, percent.include = 67)
}
\arguments{
\item{object}{An \code{ExprsArray} object to split.}

\item{percent.include}{Specifies the percent of the total number
of subjects to include in the training set (i.e., based on the larger group).
Subjects from the smaller group are up-sampled to match this number.}
}
\description{
\code{splitBoost} builds a training and validation set by randomly up-sampling
 (with replacement) the smaller of two classes. This results in an equal
 representation of each class in the training set. For example, given 30 cases and
 3 controls, a 2/3 split would place 20 cases and 20 controls in the training set.
 Of these 20 controls, only 2 are unique. The test set is not boosted. In this
 example, the test set would contain 10 cases and 1 control.
}
