% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-split.R
\name{splitBy}
\alias{splitBy}
\title{Split by User-defined Group}
\usage{
splitBy(object, colBy, include)
}
\arguments{
\item{object}{An \code{ExprsArray} object to split.}

\item{colBy}{A character string. Specifies the column used to split the data.}

\item{include}{A character vector. Specifies which annotations in \code{colBy}
to include in the training set.}
}
\value{
Returns a list of two \code{ExprsArray} objects.
}
\description{
\code{splitBy} builds a training set and validation set by placing
 all samples that have the \code{include} annotation in the specified
 \code{colBy} column in the training set. The remaining samples get
 placed in the validation set. This \code{split} is not random.
}
