% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-tidy.R
\name{validationSet}
\alias{validationSet}
\title{Extract Validation Set}
\usage{
validationSet(splitSets)
}
\value{
An \code{ExprsArray} object.
}
\description{
This function extracts the validation set from the result of a
 \code{split} method call such as \code{splitSample} or \code{splitStratify}.
}

