% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2-build.R
\name{build.}
\alias{build.}
\title{Workhorse for build Methods}
\usage{
build.(object, top, uniqueFx, ...)
}
\arguments{
\item{object}{Specifies the \code{ExprsArray} object to use as a training set
for classification.}

\item{top}{A numeric scalar or character vector. A numeric scalar indicates
the number of top features that should undergo feature selection. A character vector
indicates specifically which features by name should undergo feature selection.
Set \code{top = 0} to include all features. A numeric vector can also be used
to indicate specific features by location, similar to a character vector.}

\item{uniqueFx}{A function call unique to that fs method.}

\item{...}{Arguments passed to the respective wrapped function.}
}
\value{
Returns an \code{ExprsModel} object.
}
\description{
Used as a back-end wrapper for creating new build methods.
}

