% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.5-plNested.R
\name{check.ctrlGS}
\alias{check.ctrlGS}
\title{Check \code{ctrlGS} Arguments}
\usage{
check.ctrlGS(args)
}
\arguments{
\item{args}{A list of arguments to check.}
}
\description{
This function ensures that the list of arguments for \code{ctrlGS} meets
 the criteria required by the \code{\link{plNested}} function.
}
\details{
This function mandates \code{aucSkip = TRUE} and \code{plotSkip = TRUE}
 arguments, not unlike \code{\link{plCV}}.
}

