% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-global.R
\name{ctrlGridSearch}
\alias{ctrlGridSearch}
\title{Manage \code{plGrid} Arguments}
\usage{
ctrlGridSearch(func, top = 0, ...)
}
\arguments{
\item{func}{A character string. The \code{pl} function to call.}

\item{top}{Argument passed to the \code{pl} function. Leave missing
when handling \code{plMonteCarlo} or \code{plNested} arguments.}

\item{...}{Additional arguments passed to the \code{pl} function.}
}
\value{
A list of arguments.
}
\description{
This function organizes \code{plGrid} arguments passed to \code{pl} functions.
}
