% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.1-classes.R, R/1.2-methods.R
\docType{class}
\name{ExprsModel-class}
\alias{ExprsModel-class}
\alias{show,ExprsModel-method}
\alias{getFeatures,ExprsModel-method}
\title{An S4 class to store the model}
\usage{
\S4method{show}{ExprsModel}(object)

\S4method{getFeatures}{ExprsModel}(object)
}
\arguments{
\item{object}{An object of class \code{ExprsModel}.}
}
\description{
An S4 class to store the model
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Method to show \code{ExprsModel} object.

\item \code{getFeatures}: Method to return features within an \code{ExprsModel} object.
}}

\section{Slots}{

\describe{
\item{\code{preFilter}}{Typically a list. Stores feature selection history.}

\item{\code{reductionModel}}{Typically a list. Stores dimension reduction history.}

\item{\code{mach}}{Typically an S4 class. Stores the model.}
}}

\seealso{
\code{\link{ExprsArray-class}}\cr
\code{\link{ExprsModel-class}}\cr
\code{\link{ExprsPipeline-class}}\cr
\code{\link{ExprsEnsemble-class}}\cr
\code{\link{ExprsPredict-class}}\cr
\code{\link{RegrsPredict-class}}
}
