% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1-fs.R
\name{fsInclude}
\alias{fsInclude}
\title{Select Features by Explicit Reference}
\usage{
fsInclude(object, top = 0, include)
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo feature selection.}

\item{top}{A numeric scalar or character vector. A numeric scalar indicates
the number of top features that should undergo feature selection. A character vector
indicates specifically which features by name should undergo feature selection.
Set \code{top = 0} to include all features. A numeric vector can also be used
to indicate specific features by location, similar to a character vector.}

\item{include}{A character vector. The names of features to rank above all others.
This preserves the feature order otherwise. Argument for \code{fsInclude} only.}
}
\value{
Returns an \code{ExprsArray} object.
}
\description{
\code{fsInclude} selects features passed to the \code{include} argument.
 Ranks features by the provided order.
}
