% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1-fs.R
\name{fsRDA}
\alias{fsRDA}
\title{Reduce Dimensions by RDA}
\usage{
fsRDA(object, top = 0, colBy = NULL)
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo feature selection.}

\item{top}{A numeric scalar or character vector. A numeric scalar indicates
the number of top features that should undergo feature selection. A character vector
indicates specifically which features by name should undergo feature selection.
Set \code{top = 0} to include all features. A numeric vector can also be used
to indicate specific features by location, similar to a character vector.}

\item{colBy}{A character vector. Lists the columns in \code{@annot}
to use as the constraining matrix. Passed to \code{vegan::rda}.
Optional argument. Skip with \code{colBy = NULL}.}
}
\value{
Returns an \code{ExprsArray} object.
}
\description{
\code{fsRDA} runs an RDA using the \code{rda} function
 from the \code{vegan} package, partialling out \code{colBy}.
 The RDA model is saved and deployed automatically by the
 \code{predict} method during test set validation.
}
\details{
When \code{colBy} is provided, it serves as the constraining matrix.
 However, \code{fsRDA} always returns the unconstrained scores.
 As such, \code{fsRDA} effectively partials out the contribution
 of \code{colBy} to the training set, then uses this rule
 to partial out the contribution to the test set too.
}
