% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3-exprso.R
\name{mod}
\alias{mod}
\title{Process Data}
\description{
The \code{exprso} package includes these data process modules:

- \code{\link{modHistory}}

- \code{\link{modSubset}}

- \code{\link{modFilter}}

- \code{\link{modTransform}}

- \code{\link{modSample}}

- \code{\link{modPermute}}

- \code{\link{modInclude}}

- \code{\link{modNormalize}}

- \code{\link{modTMM}}

- \code{\link{modAcomp}}

- \code{\link{modCLR}}

- \code{\link{modRatios}}

- \code{\link{modScale}}

- \code{\link{modSkew}}
}
