% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2-modCluster.R
\docType{methods}
\name{modCluster}
\alias{modCluster}
\alias{modCluster,ExprsArray-method}
\title{Cluster Subjects}
\usage{
modCluster(object, ...)

\S4method{modCluster}{ExprsArray}(object, probes, how, onlyCluster = FALSE,
  ...)
}
\arguments{
\item{object}{An \code{ExprsArray object}. The object containing
the subject data to cluster.}

\item{...}{Additional arguments to the cluster function and/or
other functions used for clustering (e.g., \code{dist} or
\code{cutree}).}

\item{probes}{A numeric scalar or character vector. A numeric scalar indicates
the number of top features that should guide clustering. A character vector
indicates specifically which features by name should guide clustering.
Set \code{probes = 0} to include all features.}

\item{how}{A character string. The name of the function used to cluster.
Select from "hclust", "kmeans", "agnes", "clara", "diana", "fanny", or
"pam".}

\item{onlyCluster}{A logical sclar. Toggles whether to return a processed
cluster object or an updated \code{ExprsArray} object.}
}
\value{
Typically an \code{ExprsArray} object with subject cluster assignments
 added to the "cluster" column of the \code{@anot} slot.
}
\description{
This method clusters subjects based on expression data using any one of
 several available clustering algorithms. See Arguments below for details.
}
\section{Methods (by class)}{
\itemize{
\item \code{ExprsArray}: Method to compare \code{ExprsArray} objects.
}}

