% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3-cluster.R
\name{modCluster}
\alias{modCluster}
\title{Cluster Subjects}
\usage{
modCluster(object, probes, how, onlyCluster = FALSE, ...)
}
\arguments{
\item{object}{An \code{ExprsArray object}. The object containing
the subject data to cluster.}

\item{probes}{A numeric scalar or character vector. A numeric scalar indicates
the number of top features that should guide clustering. A character vector
indicates specifically which features by name should guide clustering.
Set \code{probes = 0} to include all features.}

\item{how}{A character string. The name of the function used to cluster.
Select from "hcluster", "kmeans", "agnes", "clara", "diana", "fanny", or
"pam".}

\item{onlyCluster}{A logical sclar. Toggles whether to return a processed
cluster object or an updated \code{ExprsArray} object.}

\item{...}{Additional arguments to the cluster function and/or
other functions used for clustering (e.g., \code{dist} or
\code{cutree}).}
}
\value{
Typically an \code{ExprsArray} object with subject cluster assignments
 added to the "cluster" column of the \code{@anot} slot.
}
\description{
This method clusters subjects based on expression data using any one of
 several available clustering algorithms.
}

