% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-predict.R
\docType{methods}
\name{modHistory}
\alias{modHistory}
\alias{modHistory,ExprsArray-method}
\title{Duplicate Feature Selection History}
\usage{
modHistory(object, reference)

\S4method{modHistory}{ExprsArray}(object, reference)
}
\arguments{
\item{object}{An \code{ExprsArray} object. The validation set that should undergo
feature selection and dimension reduction.}

\item{reference}{An \code{ExprsArray} or \code{ExprsModel} object. The object with
feature selection history used as a template for the validation set.}
}
\value{
Returns an \code{ExprsArray} object.
}
\description{
Duplicates the feature selection history of a reference \code{ExprsArray}
 object. Called directly by \code{predict.ExprsMachine} and indirectly by
 \code{predict.ExprsModule}. Ensures that the validation set has undergone
 the same steps of feature selection and dimension reduction as the
 training set.
}
\section{Methods (by class)}{
\itemize{
\item \code{ExprsArray}: Method to duplicate \code{ExprsArray} feature selection history.
}}

\seealso{
\code{\link{exprso-predict}}
}
