% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-mod.R
\name{modHistory}
\alias{modHistory}
\title{Replicate Data Process History}
\usage{
modHistory(object, reference)
}
\arguments{
\item{object}{An \code{ExprsArray} object. The object that should undergo a
replication of some feature selection and dimension reduction history.}

\item{reference}{An \code{ExprsArray} or \code{ExprsModel} object. The object
containing the history to use as a template.}
}
\value{
A pre-processed \code{ExprsArray} object.
}
\description{
\code{modHistory} replicates the \code{fs} history of a reference object.
 Used by \code{predict} to prepare validation set for model deployment.
}
