% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-mod.R
\name{modPermute}
\alias{modPermute}
\title{Permute Features in Data}
\usage{
modPermute(object)
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo pre-processing.}
}
\value{
A pre-processed \code{ExprsArray} object.
}
\description{
\code{modPermute} randomly samples each feature in the data
 without replacement. This method helps establish a null
 model for the purpose of testing the significance of
 observed prediction error estimates.
}
