% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9-tidy.R
\name{modSubset}
\alias{modSubset}
\title{ExprsArray Subset Wrapper}
\usage{
modSubset(object, colBy, include)
}
\description{
This function provides a tidy wrapper for the \code{ExprsArray}
 \code{subset} method.
}

