% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2-methods.R
\name{modSubset}
\alias{modSubset}
\alias{pipeSubset}
\title{Tidy Subset Wrapper}
\usage{
modSubset(object, colBy, include)

pipeSubset(object, colBy, include)
}
\arguments{
\item{object}{An \code{ExprsArray} or \code{ExprsPipeline} object to subset.}

\item{colBy}{A numeric or character index. The column that contains group annotations.}

\item{include}{A character vector. Specifies which annotations in \code{colBy}
to include in the subset.}
}
\value{
An \code{ExprsArray} or \code{ExprsPipeline} object.
}
\description{
\code{modSubset} function provides a tidy wrapper for the \code{ExprsArray}
 \code{subset} method. \code{pipeSubset} provides a tidy wrapper for the
 \code{ExprsPipeline} \code{subset} method.
}
\section{Functions}{
\itemize{
\item \code{pipeSubset}: A variant of \code{modSubset}.
}}

