% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-mod.R
\name{modTMM}
\alias{modTMM}
\title{Normalize Data}
\usage{
modTMM(object, method = "TMM")
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo pre-processing.}

\item{method}{A character string. The method used by \code{calcNormFactors}.
Defaults to the "TMM" method.}
}
\value{
A pre-processed \code{ExprsArray} object.
}
\description{
\code{modTMM} normalizes feature data.
}
\details{
This method normalizes data using the \code{calcNormFactors} function
 from the \code{edgeR} package. It returns the original counts
 multiplied by the effective library size factors.
}
