% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-mod.R
\name{modTransform}
\alias{modTransform}
\title{Log Transform Data}
\usage{
modTransform(object, base = exp(1))
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo pre-processing.}

\item{base}{A numeric scalar. The base of the logarithm.}
}
\value{
An \code{ExprsArray} object.
}
\description{
\code{modTransform} log transforms feature data.
}
