% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-mod.R
\name{modTransform}
\alias{modTransform}
\title{Log Transform Data}
\usage{
modTransform(object)
}
\arguments{
\item{object}{An \code{ExprsArray} object to undergo pre-processing.}
}
\value{
A pre-processed \code{ExprsArray} object.
}
\description{
\code{modTransform} log (base 2) transforms feature data.
}
