% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.1-pipe.R
\docType{methods}
\name{pipeUnboot}
\alias{pipeUnboot}
\alias{pipeUnboot,ExprsPipeline-method}
\title{Rename "boot" Column}
\usage{
pipeUnboot(object)

\S4method{pipeUnboot}{ExprsPipeline}(object)
}
\arguments{
\item{object}{An \code{\link{ExprsPipeline-class}} object.}
}
\value{
An \code{\link{ExprsPipeline-class}} object.
}
\description{
\code{pipeUnboot} renames the "boot" column of an \code{ExprsPipeline} object
 to "unboot".
}
\details{
This method provides a convenient adjunct to \code{\link{pipeFilter}} owing to
 how \code{exprso} handles \code{ExprsPipeline} objects with a "boot" column.
}
\section{Methods (by class)}{
\itemize{
\item \code{ExprsPipeline}: Method to rename \code{ExprsPipeline} object columns.
}}

\seealso{
\code{\link{pipeFilter}}\cr
\code{\link{pipeUnboot}}\cr
\code{\link{plCV}}\cr
\code{\link{plGrid}}\cr
\code{\link{plGridMulti}}\cr
\code{\link{plMonteCarlo}}\cr
\code{\link{plNested}}
}
