% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.3-exprso.R
\name{split}
\alias{split}
\title{Split Data}
\description{
The \code{exprso} package includes these split modules:

- \code{\link{splitSample}}

- \code{\link{splitStratify}}

- \code{\link{splitBalanced}}

- \code{\link{splitBoost}}

- \code{\link{splitBy}}
}
