% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.2-methods.R
\name{validationSet}
\alias{validationSet}
\alias{testSet}
\title{Extract Validation Set}
\usage{
validationSet(splitSets)

testSet(splitSets)
}
\arguments{
\item{splitSets}{A two-item list. The result of a \code{split} method call.}
}
\value{
An \code{ExprsArray} object.
}
\description{
This function extracts the validation set from the result of a
 \code{split} method call such as \code{splitSample} or \code{splitStratify}.
}
\section{Functions}{
\itemize{
\item \code{testSet}: A variant of \code{validationSet}.
}}

